Update sources.json
For some of the spectra without a reference, a reference on SuSpect is made to the ``Asiago SN Group''. This signifies that the data was just sent to us anytime between the date of the observation and when it appeared on SuSpect. That's a given, but I am aware a number of spectra were sent to D. Branch over the years, and we simply added these to SuSpect in later years. WISeREP simply copies this reference, so there is no better citation there. The closest there is to a bib link is the Asiago Catalog Paper (http://adsabs.harvard.edu/abs/1999A%26AS..139..531B) --> 1999A&AS..139..531B, so I have added this to the few spectra from the Asiago SN Group. 

Similarly, the same has also been done for data donated by R. Lopez. I'm guessing the one to two observations, per object, were taken and nothing was done with them, so they eventually found sanctuary at SuSpect. The nearest survey paper of R. Lopez is 1991ESOC...37..721L, and has thus been added as a reference. Note: The 90s were a different time wrt unpublished spectra. 

I have also updated a few "tmp" and "astro-ph" references. 

I imagine this issue will creep up again in future pull requests, so let's keep an eye on it and I can update/fix things here. I can also avoid renaming files that are corrected if so desired.