Delete 1998bu_19990123_3028_8492_00.dat
This spectrum is off in wavelength space by more than 40x the host redshift, and an attempt to correct to a proper observer frame reveals the wavelength solution may be off as well. I am removing this spectrum here and adding it to the list of events in need of re-acquisition (todo.md). I will submit a PR for astrocatalogs/sne later.