Rename sn1986I.p83.dat to sn1986I.p22.dat
It's unclear why superfit is placing this spectrum at +86 days post-maximum light, but Pennypacker et al. 1989 says this spectrum is a combination of June 6/9 data, or an epoch of ~ +22. I have changed the file name to reflect this, otherwise the spectrum shows up after the other at day +50 (which is correct).